pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-image-retrieval-project"
version = "0.0.1"
description = "Refactor code by LLM"
authors = [{ name = "Makoto Hirano", email = "makoto.hirano.ar@hitachi.com" }]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "base==0.0.0",
    "lmdb==1.6.2",
    "matplotlib==3.10.3",
    "numpy==2.3.1",
    "pandas==2.3.0",
    "pillow==11.3.0",
    "six==1.17.0",
    "torch==2.7.1",
    "torchvision==0.22.1",
    "tqdm==4.67.1",
]

# TODO 後で廃止予定。requirements.txtに独立パッケージを記載するようにする。
[tool.setuptools.packages.find]
where = ["src"]


[tool.black]
line-length = 88

[dependency-groups]
dev = [
    "pipreqs>=0.5.0",
]
